requirements.txt
Django>=2.0,<3.0
Django-admin-rangefilter==0.5.4
Django-allauth==0.41.0
Django-anymail==7.1.0
Django-autoslug==1.9.7
Django-braces==1.14.0
Django-ckeditor-filebrowser-filer==0.3.0
Django-cms==3.7.2
Django-cors-headers==3.2.1
Django-crispy-forms==1.9.0
Django-environ==0.4.5
Django-filer==1.7.0
Django-floppyforms==1.8.0
Django-impersonate==1.5
Django-import-export==2.0.2
Django-json-widget==1.0.1
Django-model-utils==4.0.0
Django-redis==4.11.0
Django-resized==0.3.10
Django-static-sitemaps==4.6.0
Django-storages==1.9.1
Django-tinymce==3.0.2
Djangocms-text-ckeditor==3.9.0
Djangorestframework==3.11.0
boto3==1.12.46
markdown2==2.3.8
gunicorn==20.0.4
requests==2.23.0
psycopg2>=2.7,<3.0